[TASK] Move SubstituteConstantsModeAndRequestTypeRector to v11
Resolves: #1869
