--- tests/Rector/v10/v4/SubstituteConstantsModeAndRequestType/Fixture/ext_tables.php.inc
@@ -1,15 +1,0 @@
-<?php
-
-namespace Ssch\TYPO3Rector\Tests\Rector\v10\v4\SubstituteConstantsModeAndRequestType\Fixture;
-
-defined('TYPO3_MODE') or die();
-
-?>
------
-<?php
-
-namespace Ssch\TYPO3Rector\Tests\Rector\v10\v4\SubstituteConstantsModeAndRequestType\Fixture;
-
-defined('TYPO3') or die();
-
-?>
