+++ tests/Rector/v11/v0/SubstituteConstantsModeAndRequestType/Fixture/ext_localconf.php.inc
@@ -1,6 +1,6 @@
 <?php
 
-namespace Ssch\TYPO3Rector\Tests\Rector\v10\v4\SubstituteConstantsModeAndRequestType\Fixture;
+namespace Ssch\TYPO3Rector\Tests\Rector\v11\v0\SubstituteConstantsModeAndRequestType\Fixture;
 
 defined('TYPO3_MODE') or die();
 
@@ -16,7 +16,7 @@ if(TYPO3_REQUESTTYPE_BE) {
 -----
 <?php
 
-namespace Ssch\TYPO3Rector\Tests\Rector\v10\v4\SubstituteConstantsModeAndRequestType\Fixture;
+namespace Ssch\TYPO3Rector\Tests\Rector\v11\v0\SubstituteConstantsModeAndRequestType\Fixture;
 
 defined('TYPO3') or die();
