+++ tests/Rector/v11/v0/SubstituteConstantsModeAndRequestType/Fixture/ext_tables.php.inc
@@ -1,0 +1,15 @@
+<?php
+
+namespace Ssch\TYPO3Rector\Tests\Rector\v11\v0\SubstituteConstantsModeAndRequestType\Fixture;
+
+defined('TYPO3_MODE') or die();
+
+?>
+-----
+<?php
+
+namespace Ssch\TYPO3Rector\Tests\Rector\v11\v0\SubstituteConstantsModeAndRequestType\Fixture;
+
+defined('TYPO3') or die();
+
+?>
