+++ tests/Rector/v11/v0/SubstituteConstantsModeAndRequestType/SubstituteConstantsModeAndRequestTypeRectorTest.php
@@ -2,11 +2,11 @@
 
 declare(strict_types=1);
 
-namespace Ssch\TYPO3Rector\Tests\Rector\v10\v4\SubstituteConstantsModeAndRequestType;
+namespace Ssch\TYPO3Rector\Tests\Rector\v11\v0\SubstituteConstantsModeAndRequestType;
 
 use Iterator;
 use Rector\Testing\PHPUnit\AbstractRectorTestCase;
-use Ssch\TYPO3Rector\Rector\v10\v4\SubstituteConstantsModeAndRequestTypeRector;
+use Ssch\TYPO3Rector\Rector\v11\v0\SubstituteConstantsModeAndRequestTypeRector;
 use Symplify\SmartFileSystem\SmartFileInfo;
 
 final class SubstituteConstantsModeAndRequestTypeRectorTest extends AbstractRectorTestCase
